+++ Elemental/src/com/elemental/sprite/Sprite.java
@@ -6,8 +6,8 @@ public interface Sprite {
 
 	public int getSpriteWidth();
 	public int getSpriteHeight();
-	public void setX(int x);
-	public void setY(int y);
+	public void setX(float x);
+	public void setY(float y);
 	public void draw(Canvas canvas);
 	public void update(long time);
 }
